% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckduckr-package.R
\docType{package}
\name{duckduckr}
\alias{duckduckr}
\alias{duckduckr-package}
\title{Simple Client for the DuckDuckGo Instant Answer API}
\description{
Programmatic access to the DuckDuckGo's Instant Answer API <https://api.duckduckgo.com/api>
}
\author{
Dirk Schumacher (mail@dirk-schumacher.net)
}
