% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckduck-answer.R
\name{duckduck_answer}
\alias{duckduck_answer}
\title{DuckDuckGo's Instant Answer API}
\usage{
duckduck_answer(query, no_redirect = FALSE, no_html = FALSE,
  skip_disambig = FALSE, app_name = "duckduckr")
}
\arguments{
\item{query}{the query string}

\item{no_redirect}{TRUE to skip HTTP redirects (for !bang commands)}

\item{no_html}{TRUE to remove html from results}

\item{skip_disambig}{TRUE to to skip disambiguation (D) Type.}

\item{app_name}{the appname used to identify your application.}
}
\value{
Always returns a list. If the API call was successful it contains the response of
        the duckduckgo api as parsed by \code{\link[jsonlite]{fromJSON}}. In addition the object's
        attributes contain additional meta data. Especially the status attribute indicates
        if something went wrong during the HTTP call or parsing of the json text.
}
\description{
Makes a synchronous API call to the DuckDuckGo Instant Answer API.
}
\seealso{
\url{https://api.duckduckgo.com/api} for more information on the API.
}
