% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckduck-answer.R
\name{duckduck_answer}
\alias{duckduck_answer}
\title{DuckDuckGo's Instant Answer API}
\usage{
duckduck_answer(query, no_redirect = FALSE, no_html = FALSE,
  skip_disambig = FALSE, app_name = "duckduckr")
}
\arguments{
\item{query}{the query string}

\item{no_redirect}{TRUE to skip HTTP redirects (for !bang commands)}

\item{no_html}{TRUE to remove html from results}

\item{skip_disambig}{TRUE to to skip disambiguation (D) Type.}

\item{app_name}{the appname used to identify your application. This is optional.

If the JSON cannot be parsed then this function throws an error.
If you want to be safe, try to catch those errors.}
}
\value{
Always returns a list with at least the slot `status`.
        If the status is `OK`, then the result of the query is in the `result` slot.
        If the status is `error`, then additional information is in the `error` slot.
}
\description{
Makes a synchronous API call.
}
